// src/store/procuring/type.ts
export interface ProcuringFormData {
  id:string;
  title: string;
  user_id: string;
  material_id: string;
  coil_count: number;
  material_unit_id: string;
  unit_value: number;
  price: number;
  currency_id: string;
  currency_value: number;
  spent_count: number;
  receipts_photo:string;
  notes: string;
  date: string;
  approved_by: string;
  check_date: string;
}